Use immutable storage for ConstantWeightedProduct validation (#49)
* add constant weighted invariant and constant sum/prod invariant, this last one with two different implementations

* add validateOutGivenIn and make validation functions non-view

* update strategies to include pair balance validation and all balances validation, added some temporal tests

* create separate strategies interfaces and added poolId param to validation functions

* use immutable storage for ConstantWeightedProduct validation, Closes #45

* packed correctly weights, added constructor verifications and tests to check everything is consistent

* refactor shiftWeights and make it a bit more clear

* rename validateAll to validateTuple in TupleTradingStrategy

* fix outdated comments

* make weights and totalTokens private

* update comments

* remove weight scaling factor

+++ contracts/strategies/IPairTradingStrategy.sol
@@ -17,8 +17,8 @@ pragma solidity ^0.7.1;
 interface IPairTradingStrategy {
     function validatePair(
         bytes32 poolId,
-        uint256 tokenIndexIn,
-        uint256 tokenIndexOut,
+        uint8 tokenIndexIn,
+        uint8 tokenIndexOut,
         uint256 tokenBalanceIn,
         uint256 tokenBalanceOut,
         uint256 tokenAmountIn,
